expr_t coerce doesn't care about poor uint64_t -> double data loss.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16620 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/simulationcraft.hpp
@@ -2151,7 +2151,7 @@ class expr_t
   std::string name_;
 
 protected:
-  template <typename T> static double coerce( T t ) { return t; }
+  template <typename T> static double coerce( T t ) { return static_cast<double>( t ); }
   static double coerce( timespan_t t ) { return t.total_seconds(); }
 
   virtual double evaluate() = 0;
